test: add native shim unit tests and fix regex/okhttp semantics
- jsoup: 10 tests via real dom_query documents (select, text, attrs,
  children/html, Response seam, empty selection)
- java.util.regex: 8 tests (matches/find/group/split/quote/reset);
  fix matcher_end returning start, split(limit=0) applying no splits,
  and Pattern.matches/Kotlin Regex.matches requiring full-string match
- kotlin: 9 tests (Pair, collections, startsWith$default mask,
  Regex full-match/replace, Duration conversions)
- okhttp: 9 tests (Request builder chain, headers, Response accessors,
  HttpUrl parsing, interceptor lists); fix pathSegments including host
- android + injekt: smoke tests
- gate apk-based context tests on tachiyomi+okhttp; drop dead HTTP_URL
- clippy fixes in dexcli/dump

+++ examples/dump.rs
@@ -5,8 +5,8 @@ fn main() {
     let dex = DexFile::parse(&data).unwrap();
     for cd in &dex.classes {
         let Some(d) = &cd.class_data else { continue };
-        let desc = dex.type_descriptor(cd.class_idx as u32).to_string();
-        let mk = |m: &dexvm::dex::EncodedMethod| dex.method_key(m.method_idx as u32).unwrap();
+        let desc = dex.type_descriptor(cd.class_idx).to_string();
+        let mk = |m: &dexvm::dex::EncodedMethod| dex.method_key(m.method_idx).unwrap();
         let has = d
             .direct_methods
             .iter()
